Add SSH key deletion support. Add orphaned resources check. (#9)
* Add SSH key deletion support. Add orphaned resources check.

* Fix linter name and flag name.

* Add testing.

* Ensure at compile time that types implement interfaces.

* Allow many methods on interface.

+++ internal/civo/volumes.go
@@ -1,4 +1,3 @@
-//nolint:dupl // similar functions due to upstream packaging
 package civo
 
 import (
@@ -7,6 +6,8 @@ import (
 	"github.com/konstructio/dropkick/internal/compare"
 	"github.com/konstructio/dropkick/internal/outputwriter"
 )
+
+const volumeStatusAttached = "attached"
 
 // NukeVolumes deletes all volumes associated with the Civo client.
 // It returns an error if the deletion process encounters any issues.
@@ -21,7 +22,7 @@ func (c *Civo) NukeVolumes() error {
 	c.logger.Infof("found %d volumes", len(volumes))
 
 	for _, volume := range volumes {
-		c.logger.Infof("found volume %q - ID: %q", volume.Name, volume.ID)
+		c.logger.Infof("found volume %q - ID: %q (attached? %v)", volume.Name, volume.ID, volume.Status == volumeStatusAttached)
 
 		if c.nameFilter != "" && !compare.ContainsIgnoreCase(volume.Name, c.nameFilter) {
 			c.logger.Warnf("skipping volume %q: name does not match filter", volume.Name)
@@ -29,6 +30,11 @@ func (c *Civo) NukeVolumes() error {
 		}
 
 		if c.nuke {
+			if volume.Status == volumeStatusAttached {
+				c.logger.Warnf("refusing to delete volume %s: it is attached to the node instance with ID %q", volume.Name, volume.InstanceID)
+				continue
+			}
+
 			c.logger.Infof("deleting volume %q", volume.Name)
 
 			res, err := c.client.DeleteVolume(volume.ID)
